add machines backend, Production Line and schedule updates

+++ apps/api/src/production-lines/production-lines.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { EntityManager, Repository } from 'typeorm';
 import { FinalGoodsService } from '../final-goods/final-goods.service';
 import { OrganisationsService } from '../organisations/organisations.service';
+import { Schedule } from '../schedules/entities/schedule.entity';
 import { CreateProductionLineDto } from './dto/create-production-line.dto';
 import { UpdateProductionLineDto } from './dto/update-production-line.dto';
 import { ProductionLine } from './entities/production-line.entity';
@@ -45,7 +46,8 @@ export class ProductionLinesService {
       relations: {
         finalGood: true,
         schedules: true,
-        organisation: true
+        organisation: true,
+        machines: true
       }
     })
   }
@@ -56,7 +58,8 @@ export class ProductionLinesService {
     }, relations: {
       finalGood: true,
       schedules: true,
-      organisation: true
+      organisation: true,
+      machines: true
     }})
   }
 
@@ -66,6 +69,7 @@ export class ProductionLinesService {
     }, relations: {
       finalGood: true,
       schedules: true,
+      machines: true,
     }})
   }
 
@@ -73,16 +77,21 @@ export class ProductionLinesService {
     const productionLineToUpdate = await this.findOne(id)
     const keyValuePairs = Object.entries(updateProductionLineDto)
     keyValuePairs.forEach(([key, value]) => {
-      if (value) {
-        productionLineToUpdate[key] = value
-      }
+      productionLineToUpdate[key] = value
     })
     return this.productionLineRepository.save(productionLineToUpdate)
   }
 
   async remove(id: number): Promise<ProductionLine> {
-    const productionLineToRemove = await this.productionLineRepository.findOneBy({id})
-    return this.productionLineRepository.remove(productionLineToRemove)
+    //only can remove if all schedules are completed
+    const productionLineToRemove = await this.findOne(id)
+    const ongoingSchedules = productionLineToRemove.schedules.some(schedule => schedule.status === 'ongoing')
+    if (!ongoingSchedules) {
+      return this.productionLineRepository.remove(productionLineToRemove)
+    } else {
+      throw new NotFoundException('error deleting production line as there is an ongoing schedule')
+    }
+    
   }
 
   async updateNextAvailable(id: number, newScheduleEndDate: Date, entityManager: EntityManager) {
@@ -93,5 +102,51 @@ export class ProductionLinesService {
     const nextAvailableDTInMilliseconds = endDateInMilliseconds + productionLine.changeOverTime
     const nextAvailableDTInDateObject = new Date(nextAvailableDTInMilliseconds)
     await entityManager.update(ProductionLine, productionLine.id, {nextAvailableDateTime: nextAvailableDTInDateObject})
+  }
+
+  async machineTriggerChange(machineIsOperating: Boolean, machineId: number, productionLineId: number, entityManager: EntityManager) {
+    //if its true, check if the status of productionLine is not available 
+    //check if there are other machines that are false, 
+    //If there are, do nothing
+    //if this is the only one, update status of PL to true and update all schdules with the time elapsed 
+    const productionLine = await entityManager.findOne(ProductionLine, {
+      where: {
+        id: productionLineId
+      },
+      relations: {
+        schedules: true,
+        machines: true
+      }
+    })
+    const schedules = productionLine.schedules
+    const machines = productionLine.machines
+    if (machineIsOperating) {
+      if (!productionLine.isAvailable) {
+        const machinesNotInOperation = machines.filter(machine => !machine.isOperating)
+        if (machinesNotInOperation.length === 1 && machinesNotInOperation[0].id === machineId) {
+          const timeElapsedInMilliseconds = new Date().getTime() - productionLine.lastStopped.getTime()
+          //update ongoing schedule
+          const ongoingSchedule = schedules.filter(schedule => schedule.status === 'ongoing')
+          if (ongoingSchedule.length === 1) {
+            const schedule = ongoingSchedule[0]
+            const newEndTime = new Date(schedule.end).getTime() + timeElapsedInMilliseconds
+            await entityManager.update(Schedule, schedule.id, {end: new Date(newEndTime)})
+          } 
+          //update nextAvailableDateTime
+          const newNextAvailableDateTime = new Date(productionLine.nextAvailableDateTime).getTime() + timeElapsedInMilliseconds
+          productionLine.nextAvailableDateTime =  new Date(newNextAvailableDateTime)
+          productionLine.lastStopped = null
+          productionLine.isAvailable = true
+          await entityManager.save(ProductionLine, productionLine)
+        }
+      }
+    } else {
+        //if value is false
+        if (productionLine.isAvailable) {
+          productionLine.isAvailable = false
+          productionLine.lastStopped = new Date()
+          await entityManager.save(productionLine)
+        }
+    }
   }
 }
